fix(layout): fix layout convert error

+++ layout/engine/HPUtil.cpp
@@ -103,7 +103,7 @@ float HPRoundValueToPixelGrid(float value, float scaleFactor, bool forceCeil, bo
     // First we check if the value is already rounded
     scaleValue = scaleValue - fractial;
   } else if (FloatIsEqual(fractial, 1.0)) {
-    scaleValue = static_cast<float>(scaleValue - fractial + 1.0);
+    scaleValue = scaleValue - fractial + 1.0f;
   } else if (forceCeil) {
     // Next we check if we need to use forced rounding
     scaleValue = scaleValue - fractial + 1.0f;
